refactor conversion to journal
