+++ crates/proof/Cargo.toml
@@ -29,7 +29,7 @@ tracing.workspace = true
 cfg-if.workspace = true
 thiserror.workspace = true
 
-canoe-bindings = { workspace = true, optional = true }
+canoe-bindings = { workspace = true }
 
 # Steel 
 canoe-steel-methods = { workspace = true, optional = true }
@@ -40,5 +40,5 @@ sp1-lib = { workspace = true, features = ["verify"], optional = true}
 sha2 = { workspace = true, optional = true }
 
 [features]
-steel = ["canoe-steel-methods", "risc0-zkvm", "canoe-bindings"]
-sp1-cc = ["canoe-bindings", "sha2", "sp1-lib"]
+steel = ["canoe-steel-methods", "risc0-zkvm"]
+sp1-cc = ["sha2", "sp1-lib"]
