+++ crates/proof/src/canoe_verifier/sp1_cc.rs
@@ -1,12 +1,7 @@
-use crate::canoe_verifier::{CanoeVerifier, VERIFIER_ADDRESS};
+use crate::canoe_verifier::CanoeVerifier;
 use crate::cert_validity::CertValidity;
-
-use eigenda_v2_struct::EigenDAV2Cert;
-
-use alloc::vec::Vec;
 use alloy_primitives::B256;
-use alloy_sol_types::SolValue;
-use canoe_bindings::Journal;
+use eigenda_v2_struct::EigenDAV2Cert;
 
 use tracing::{info, warn};
 
@@ -19,40 +14,20 @@ pub const VKEYHEXSTRING: &str = "001a1106242f4bf2a44b02aeb0123dec8a842654b3cf941
 pub struct CanoeSp1CCVerifier {}
 
 impl CanoeVerifier for CanoeSp1CCVerifier {
+    #[allow(unused_variables)]
     fn validate_cert_receipt(&self, cert_validity: CertValidity, eigenda_cert: EigenDAV2Cert) {
         info!("using CanoeSp1CCVerifier");
-        // if not in dev mode, the receipt must be non empty
-
-        let batch_header = eigenda_cert.batch_header_v2.to_sol().abi_encode();
-        let blob_inclusion_info = eigenda_cert.blob_inclusion_info.to_sol().abi_encode();
-        let non_signer_stakes_and_signature = eigenda_cert
-            .nonsigner_stake_and_signature
-            .to_sol()
-            .abi_encode();
-        let signed_quorum_numbers_abi = eigenda_cert.signed_quorum_numbers.abi_encode();
-
-        // ensure inputs are constrained
-        let mut buffer = Vec::new();
-        buffer.extend(batch_header);
-        buffer.extend(blob_inclusion_info);
-        buffer.extend(non_signer_stakes_and_signature);
-        buffer.extend(signed_quorum_numbers_abi);
-
-        let journal = Journal {
-            certVerifierAddress: VERIFIER_ADDRESS,
-            input: buffer.into(),
-            blockhash: cert_validity.l1_head_block_hash,
-            output: cert_validity.claimed_validity,
-            l1ChainId: cert_validity.l1_chain_id,
-        };
-        let journal_bytes = journal.abi_encode();
 
         cfg_if::cfg_if! {
             if #[cfg(target_os = "zkvm")] {
                 use sha2::{Digest, Sha256};
                 use sp1_lib::verify::verify_sp1_proof;
                 use core::str::FromStr;
+                use crate::canoe_verifier::to_journal_bytes;
 
+                let journal_bytes = to_journal_bytes(&cert_validity, &eigenda_cert);
+
+                // if not in dev mode, the receipt should be empty
                 if cert_validity.canoe_proof.is_some() {
                     // Sp1 doc https://github.com/succinctlabs/sp1/blob/a1d873f10c32f5065de120d555cfb53de4003da3/examples/aggregation/script/src/main.rs#L75
                     warn!("sp1-cc verification within zkvm requires proof being provided via zkVM stdin");
@@ -64,7 +39,6 @@ impl CanoeVerifier for CanoeSp1CCVerifier {
                 verify_sp1_proof(&v_key, &public_values_digest.into());
             } else {
                 warn!("Sp1CC proof IS NOT verified in the non zkVM environment");
-                _ = journal_bytes;
             }
         }
     }
