+++ crates/proof/src/canoe_verifier/steel.rs
@@ -1,8 +1,5 @@
-use crate::canoe_verifier::{CanoeVerifier, VERIFIER_ADDRESS};
+use crate::canoe_verifier::{to_journal_bytes, CanoeVerifier};
 use crate::cert_validity::CertValidity;
-use alloc::vec::Vec;
-use alloy_sol_types::SolValue;
-use canoe_bindings::Journal;
 use eigenda_v2_struct::EigenDAV2Cert;
 
 use risc0_zkvm::Receipt;
@@ -23,28 +20,7 @@ impl CanoeVerifier for CanoeSteelVerifier {
     fn validate_cert_receipt(&self, cert_validity: CertValidity, eigenda_cert: EigenDAV2Cert) {
         info!("using CanoeSteelVerifier");
 
-        let batch_header = eigenda_cert.batch_header_v2.to_sol().abi_encode();
-        let blob_inclusion_info = eigenda_cert.blob_inclusion_info.to_sol().abi_encode();
-        let non_signer_stakes_and_signature = eigenda_cert
-            .nonsigner_stake_and_signature
-            .to_sol()
-            .abi_encode();
-        let signed_quorum_numbers_abi = eigenda_cert.signed_quorum_numbers.abi_encode();
-
-        let mut buffer = Vec::new();
-        buffer.extend(batch_header);
-        buffer.extend(blob_inclusion_info);
-        buffer.extend(non_signer_stakes_and_signature);
-        buffer.extend(signed_quorum_numbers_abi);
-
-        let journal = Journal {
-            certVerifierAddress: VERIFIER_ADDRESS,
-            input: buffer.into(),
-            blockhash: cert_validity.l1_head_block_hash,
-            output: cert_validity.claimed_validity,
-            l1ChainId: cert_validity.l1_chain_id,
-        };
-        let journal_bytes = journal.abi_encode();
+        let journal_bytes = to_journal_bytes(&cert_validity, &eigenda_cert);
 
         cfg_if::cfg_if! {
             if #[cfg(target_os = "zkvm")] {
